PgfDB::ref_count must be size_t

+++ src/runtime/c/pgf/db.h
@@ -67,7 +67,7 @@ private:
 public:
     // Here we count to how many revisions the client has access.
     // When the count is zero we release the database.
-    int ref_count;
+    size_t ref_count;
 
     PGF_INTERNAL_DECL PgfDB(const char* filepath, int flags, int mode);
     PGF_INTERNAL_DECL ~PgfDB();
